+++ post.ts
@@ -1,29 +1,5 @@
-import * as cache from '@actions/cache';
 import * as core from '@actions/core';
-import * as exec from '@actions/exec';
-import { CACHE_ENABLED, getPathsToCache, getPrimaryCacheKey } from './helpers';
-
-async function saveCache() {
-	if (!CACHE_ENABLED) {
-		return;
-	}
-
-	const primaryKey = getPrimaryCacheKey();
-	const cacheHitKey = core.getState('cache-hit-key');
-
-	if (cacheHitKey === primaryKey) {
-		core.info(`Cache hit occured on the key ${cacheHitKey}, not saving cache`);
-		return;
-	}
-
-	core.info('Cleaning cache before saving');
-
-	await exec.exec('cargo', ['cache', '--autoclean']);
-
-	core.info(`Saving cache with key ${primaryKey}`);
-
-	await cache.saveCache(getPathsToCache(), primaryKey);
-}
+import { saveCache } from './src/cargo';
 
 async function run() {
 	try {
